- Replace enter pin

+++ stepdefinitions/ui/smartforms/oily-tanker/section5-steps.rb
@@ -1,15 +1,11 @@
 # frozen_string_literal: true
 
-And (/^I (select|delete) (.+) role from list$/) do |_condition, _roles|
-  if _condition === 'select'
-    on(Section5Page).select_roles_and_responsibility(_roles)
-  end
-  if _condition === 'delete'
-    on(Section5Page).delete_roles_and_responsibility(_roles)
-  end
+And(/^I (select|delete) (.+) role from list$/) do |_condition, _roles|
+  on(Section5Page).select_roles_and_responsibility(_roles) if _condition === 'select'
+  on(Section5Page).delete_roles_and_responsibility(_roles) if _condition === 'delete'
 end
 
-And (/^I fill up section 5$/) do
+And(/^I fill up section 5$/) do
   step 'I select 1 role from list'
   step 'I fill up non crew details'
   step 'I sign on role with sponsor crew 8383 pin'
@@ -20,17 +16,17 @@ And (/^I fill up section 5$/) do
   # step 'I select 1 role from list'
 end
 
-And (/^I sign on listed role$/) do
+And(/^I sign on listed role$/) do
   BrowserActions.scroll_click(on(Section5Page).sign_btn_role_elements.first)
-  step "I sign on canvas with valid A/M rank"
+  step 'I sign on canvas with valid A/M rank'
   sleep 1
 end
 
-Then (/^I should see (.+) role listed$/) do |_total_roles|
+Then(/^I should see (.+) role listed$/) do |_total_roles|
   is_equal(on(Section5Page).responsibility_box_elements.size, _total_roles)
 end
 
-And (/^I (should|should not) see (.+) role$/) do |_condition, _role|
+And(/^I (should|should not) see (.+) role$/) do |_condition, _role|
   sleep 1
   if _condition === 'should'
     is_true(on(Section5Page).is_role?(_role))
@@ -39,48 +35,48 @@ And (/^I (should|should not) see (.+) role$/) do |_condition, _role|
   end
 end
 
-When (/^I delete the role from cross$/) do
+When(/^I delete the role from cross$/) do
   on(Section3CPage).cross_btn_elements.first.click
 end
 
-Then (/^I should see a list of roles$/) do
+Then(/^I should see a list of roles$/) do
   on(Section5Page).roles_and_resp_btn
   is_true(on(Section5Page).is_list_of_roles?)
 end
 
-And (/^I sign on role with non sponsor crew (.*) rank$/) do |_rank|
+And(/^I sign on role with non sponsor crew (.*) rank$/) do |_rank|
   on(Section5Page).sign_btn_role_elements.first.click
   step "I enter pin for rank #{_rank}"
 end
 
-And (/^I sign on role$/) do
+And(/^I sign on role$/) do
   on(Section5Page).sign_btn_role_elements.first.click
-  step "I sign on canvas with valid A/M rank for fsu"
+  step 'I sign on canvas with valid A/M rank for fsu'
   step 'I set time'
 end
 
-And (/^I sign on next role with same user$/) do
+And(/^I sign on next role with same user$/) do
   on(Section5Page).sign_btn_role_elements.last.click
-  step "I sign on canvas with valid A/M rank for fsu"
+  step 'I sign on canvas with valid A/M rank for fsu'
   step 'I set time'
 end
 
-Then (/^I should see (.*) signed role details with (.*) pin$/) do |_which_role,_pin|
-  is_true(on(Section5Page).is_role_signed_user_details?(_which_role,_pin))
+Then(/^I should see (.*) signed role details with (.*) pin$/) do |_which_role, _pin|
+  is_true(on(Section5Page).is_role_signed_user_details?(_which_role, _pin))
 end
 
-And (/^I fill up non crew details$/) do
+And(/^I fill up non crew details$/) do
   step 'I check non crew member checkbox'
   on(Section5Page).other_name = 'Test Automation'
   on(Section5Page).other_company = 'Test Automation Company'
 end
 
-Then (/^I should see non crew copy text$/) do
+Then(/^I should see non crew copy text$/) do
   sleep 1
   is_equal(on(Section5Page).non_crew_copy_text, 'Ship Staff to use PIN for non-crew member to enter signature')
 end
 
-And (/^I check non crew member checkbox$/) do
+And(/^I check non crew member checkbox$/) do
   on(Section5Page).non_crew_checkbox_elements.first.click
 end
 
@@ -93,24 +89,24 @@ end
 #   end
 # end
 
-And (/^I sign on role with sponsor crew (.+) pin$/) do |_pin|
+And(/^I sign on role with sponsor crew (.+) rank$/) do |_rank|
   on(Section5Page).sign_btn_role_elements.first.click
-  step "I sign on canvas with valid #{_pin} pin for fsu"
+  step "I sign on canvas with valid #{_rank} rank for fsu"
   step 'I set time'
 end
 
-Then (/^I should see non crew details$/) do
+Then(/^I should see non crew details$/) do
   sleep 1
   is_equal(on(Section5Page).signed_rank_and_name_elements.first.text, 'Test Automation')
   if on(Section5Page).get_non_crew_date_time_element.text === "#{on(Section5Page).get_current_date_format_with_offset} #{on(Section5Page).get_current_time_format}"
-    is_equal(on(Section5Page).get_non_crew_date_time_element.text, "#{on(Section5Page).get_current_date_format_with_offset} #{on(Section5Page).get_current_time_format}")
+    is_equal(on(Section5Page).get_non_crew_date_time_element.text,
+             "#{on(Section5Page).get_current_date_format_with_offset} #{on(Section5Page).get_current_time_format}")
   else
-    is_equal(on(Section5Page).get_non_crew_date_time_element.text, "get_current_date_and_time_minus_a_min")
+    is_equal(on(Section5Page).get_non_crew_date_time_element.text, 'get_current_date_and_time_minus_a_min')
   end
-  
 end
 
-And (/^I should see supervise by (.+) detail and (.+) detail$/) do |_supervized, _company|
+And(/^I should see supervise by (.+) detail and (.+) detail$/) do |_supervized, _company|
   is_equal(on(Section5Page).get_filled_crew_details_elements.first.text, _supervized)
   is_equal(on(Section5Page).get_filled_crew_details_elements.last.text, _company)
 end
